Remove unnecessary type coversions

+++ mounts_darwin.go
@@ -77,13 +77,13 @@ func mounts() ([]Mount, []string, error) {
 			Type:       fsType,
 			Opts:       opts,
 			Metadata:   stat,
-			Total:      (uint64(stat.Blocks) * uint64(stat.Bsize)),
-			Free:       (uint64(stat.Bavail) * uint64(stat.Bsize)),
-			Used:       (uint64(stat.Blocks) - uint64(stat.Bfree)) * uint64(stat.Bsize),
+			Total:      stat.Blocks * uint64(stat.Bsize),
+			Free:       stat.Bavail * uint64(stat.Bsize),
+			Used:       (stat.Blocks - stat.Bfree) * uint64(stat.Bsize),
 			Inodes:     stat.Files,
 			InodesFree: stat.Ffree,
 			InodesUsed: stat.Files - stat.Ffree,
-			Blocks:     uint64(stat.Blocks),
+			Blocks:     stat.Blocks,
 			BlockSize:  uint64(stat.Bsize),
 		}
 		d.DeviceType = deviceType(d)
